fe: remove deprecated code
