+++ frontend/src/components/index.ts
@@ -5,7 +5,5 @@ export * from "./LiveResults";
 export * from "./CurrentPhase";
 export * from "./Quorum";
 export * from "./TopVoters";
-export * from "./ValidatorDetail";
-export * from "./VotesBubbleChart";
 export * from "./Footer";
 export * from "./VotingProposalsDropdown";
