--- frontend/src/data/OLD/getAllProposals.ts
@@ -1,10 +1,0 @@
-import { program } from "@/chain/helpers";
-
-/**
- * @deprecated old program
- */
-export const getAllProposals = async () => {
-  const proposals = await program.account.proposal.all();
-
-  return proposals;
-};
